+++ CHANGELOG.md
@@ -3,16 +3,17 @@
 All notable changes to this project will be documented in this file.
 
 ###
-## [v1.0.1] / 2026-xx-xx
+## [v1.1.0] / 2026-xx-xx
 
 ### Added
 - Plugin window can be maximized/minimized
 - Settings window can be resized
 - Checksum verificator in "Manual Input" tab
 - New option in settings to keep sending data when window is closed
-- Button to copy sentence in "Manual Input" tab
+- Button to copy to clipboard the sentence in "Manual Input" tab
 - New option to automatically copy NMEA sentence to clipboard when manually sending in "Sentence builder"
 - Configurable timer for sending simulation data
+- Dual vessel simulation: Own ship and AIS target can now run simultaneously
 
 ### Fixed
 - Sentences with "!" as first character are accepted.